Refactor train.py file into multiple files
Refactor the `train.py` file into multiple files for better modularity and organization.

* **Data Preparation**:
  - Add `train/data_preparation.py` to define functions `load_and_prepare_data` and `create_timeseries_generator`.
  - Import necessary libraries and modules.

* **Model Building**:
  - Add `train/model.py` to define the function `build_model`.
  - Import necessary libraries and modules.

* **Training**:
  - Add `train/training.py` to define functions `train_model`, `calculate_metrics`, and `save_and_log_plots`.
  - Import necessary libraries and modules.

* **Optuna Tuning**:
  - Add `train/optuna_tuning.py` to define functions `objective` and `perform_hyperparameter_tuning`.
  - Import necessary libraries and modules.

* **Main Script**:
  - Modify `train/train.py` to remove data preparation, model building, training, and Optuna tuning code.
  - Import functions from `data_preparation.py`, `model.py`, `training.py`, and `optuna_tuning.py`.
  - Update the main function to use the imported functions.
